+++ src/entities/FarmerPlotEntities.ts
@@ -7,7 +7,7 @@ import { LAT_LNG_BANGKOK } from "../definitions/Location";
 export interface FarmerPlotEntity {
   id?: string;
   plotName: string;
-  raiAmount: number;
+  raiAmount?: string;
   landmark: string;
   plantName: string;
   plantNature: string;
@@ -20,11 +20,12 @@ export interface FarmerPlotEntity {
   plotId: number;
   farmerId?: string;
   plotAreaId: number;
+  comment?: string;
 }
 export const FarmerPlotEntity_INIT: FarmerPlotEntity = {
   id: "",
   plotName: "",
-  raiAmount: 0,
+  raiAmount: "",
   landmark: "",
   plantName: "",
   plantNature: "",